Move view-related components into view/
